Changes in stacks
- aws: added calls in rpc.go
- aws: added handling of aws.Error with code "InvalidParameterValue"
- aws: changes Security Group rule convert to AWS counterpart to support use of Security Group ID as source/target, supplementary to IP ranges (but no mix allowed and checked)
- all: Security Group rule port type replaced from uint16 to int32 (may contain -1 meaning all ports)
- all: consolidated abstract.HostFull.CurrentState values
+ code cleanup

+++ lib/server/iaas/stacks/openstack/securitygroup.go
@@ -296,8 +296,8 @@ func convertRulesToAbstract(in []secrules.SecGroupRule) ([]abstract.SecurityGrou
 			Description: v.Description,
 			Direction:   direction,
 			Protocol:    v.Protocol,
-			PortFrom:    uint16(v.PortRangeMin),
-			PortTo:      uint16(v.PortRangeMax),
+			PortFrom:    int32(v.PortRangeMin),
+			PortTo:      int32(v.PortRangeMax),
 			IPRanges:    []string{v.RemoteIPPrefix},
 		}
 		out = append(out, n)
